+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Added:
 Fixes:
 
 * Improved the logic of "fastest lap across multiple results files" sorting in Race Weekends whilst using fallback results sorting (premium)
+* Fixes an issue where Championship Driver and Team penalties would be lost when editing a Championship.
 * Drastically improved the speed of STracker pages through some reworking of how STracker is proxied.
 * Fixed an issue where some STracker links would not work correctly.